Add type declaration file
Hand roll a type declaration file for `flat.js` that accurately reflects
the exported plugin. This resolves type errors when attempting to use
`globals` and `recommended` configs in type-checked eslint configs and
also aids in autocomplete when typing (as in keyboard button press, not
"does it satisfy expected type") properties of the plugin.